+++ packages/web3-providers/src/CoinMarketCap/apis/CoinMarketCapSearchAPI.ts
@@ -1,8 +1,8 @@
 import urlcat from 'urlcat'
 import type { NonFungibleTokenResult, SearchResult, SourceType } from '@masknet/web3-shared-base'
-import { fetchJSON } from '../helpers/fetchJSON.js'
-import type { DSearchBaseAPI } from '../types/DSearch.js'
-import { DSEARCH_BASE_URL } from '../DSearch/constants.js'
+import { fetchJSON } from '../../helpers/fetchJSON.js'
+import type { DSearchBaseAPI } from '../../types/DSearch.js'
+import { DSEARCH_BASE_URL } from '../../DSearch/constants.js'
 
 export interface FungibleToken {
     id: string | number
